combat: long/per-mille magnitude boundary + checked overflow for overlay damage
OverlayCombatCalculator/ElementHub/EntityStatWriter (lawn-combat-wire T4, combat-numerics):

- OverlayCombatRequest.BaseOverlayDamage is now long, not double -- every real caller
  already handed it an integral value.
- The parry/block neutral-share computation now multiplies by the per-mille share
  before dividing by 1000, instead of dividing first (the named divide-before-multiply
  defect).
- Every long-narrowing cast on the magnitude path (parry/block chip removal, the final
  signed delta, OverlayCombatMath's heal narrowing) is wrapped in checked, so an
  out-of-range magnitude throws instead of silently narrowing.
- ElementHub's STR/WEK/NEU multiplier is composed in long per-mille internally, with a
  single division at the public double boundary IElementHub still requires.
- EntityStatWriter's Unity int32 narrowing now goes through one reporting wrapper that
  logs and emits a proof event when a magnitude actually exceeds int32, instead of
  saturating silently; documented as the structural host-field-width boundary it is.

The continuous mitigation-chain interior (power/defense/pierce/amp/crit/sigmoid) stays
double on purpose -- CombatDerivedReader/CombatPolicy/CombatProbability and
IElementHub.cs are out-of-scope, already-double-typed dependencies (see each touched
file's own class-doc comment for the citation trail), and Actions/BasicAttack.cs has a
live production dependency on EffectivenessMultiplier's double shape.

Added OverlayCombatNumericsTests.cs: overflow-throws, divide-last (numeric +
structural), a direct MergeAppliedCombat combat-channel guard, and a documented
double/float allowlist per file. Updated four existing test helpers whose double
default parameters no longer match BaseOverlayDamage's new long type.

dotnet test tests/FusionRpg.Core.Tests --filter "FullyQualifiedName~OverlayCombat|ElementHub":
94/94 passed. scripts/audit-overflow.py on Combat+Injector/Stats: 0 critical (A3 now 0,
was 1 pre-existing; A7's 12 double findings are the documented, out-of-scope exception).

+++ tests/FusionRpg.Core.Tests/Combat/MinChipFloorTests.cs
@@ -8,7 +8,7 @@ namespace FusionRpg.Core.Tests.Combat;
 /// <summary>U4 — per-profile min-chip floor (owner decision 6).</summary>
 public class MinChipFloorTests
 {
-    static OverlayCombatRequest TankMatchup(CombatProfile profile, double baseDamage = 100,
+    static OverlayCombatRequest TankMatchup(CombatProfile profile, long baseDamage = 100,
         bool? forceHit = true, bool? forceCrit = false)
     {
         var composer = new DerivedComposer();
